Added NIHR data processing
